requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.26.0
apscheduler==3.10.4
stripe==7.11.0
jinja2==3.1.3
aiosmtplib==3.0.1
email-validator==2.1.0
arq==0.25.0
redis==5.0.1
arq==0.25.0
redis==5.0.1
arq==0.25.0
redis==5.0.1
